requirements.txt
aiohttp==3.7.4
pytest==7.2.4
pytest-aiohttp==0.6.0
pytest-cov==3.0.0
allure-pytest==2.9.36